+++ spec/luacheck_spec.lua
@@ -499,53 +499,6 @@ describe("process_reports", function()
          std = "none"
       })))
    end)
-
-   -- There is no place to test standards.
-   it("works with tarantool std", function()
-      local source = [[
-            box.backup.start()
-            box.schema.space.create()
-            package.setsearchroot()
-            box.backup.invalid()
-            package.invalid()
-]]
-      assert.same({
-         {
-            {code = "143", name = "box", field = "backup.invalid", indexing = {"backup", "invalid"}},
-            {code = "143", name = "package", field = "invalid", indexing = {"invalid"}},
-         },
-         warnings = 2,
-         errors = 0,
-         fatals = 0
-      }, strip_locations(luacheck.process_reports({luacheck.get_report(source)}, {std = "tarantool"})))
-      assert.same({
-         {
-            {code = "113", name = "box", indexing = {"backup", "start"}},
-            {code = "113", name = "box", indexing = {"schema", "space", "create"}},
-            {code = "143", name = "package", field = "setsearchroot", indexing = {"setsearchroot"}},
-            {code = "113", name = "box", indexing = {"backup", "invalid"}},
-            {code = "143", name = "package", field = "invalid", indexing = {"invalid"}},
-         },
-         warnings = 5,
-         errors = 0,
-         fatals = 0
-      }, strip_locations(luacheck.process_reports({luacheck.get_report(source)}, {std = "max"})))
-
-      assert.same({
-         {
-            {code = "143", name = "box", field = "session.invalid", indexing = {"session", "invalid"}},
-         },
-         warnings = 1,
-         errors = 0,
-         fatals = 0,
-      }, strip_locations(luacheck.process_reports({luacheck.get_report([[
-         box.session.storage.field1 = 1
-         box.session.storage['field11'] = 1
-         print(box.session.storage.field2, box.session.storage['field22'])
-         box.session.push()
-         box.session.invalid()
-]])}, {std = "tarantool"})))
-   end)
 end)
 
 describe("get_message", function()